fixed by mkurnikov recommendations
