+++ dove/tests/cmd_dove_run_with_args.rs
@@ -1,0 +1,165 @@
+#![cfg(test)]
+
+use std::path::{Path, PathBuf};
+use std::process::{Command};
+use std::fs::{remove_dir_all, read_to_string};
+use fs_extra::file::write_all;
+
+/// $ dove run rdemo.move
+/// project name: demoproject_8
+#[test]
+fn with_args() {
+    // Path to dove folder
+    let dove_folder = {
+        let mut folder = Path::new(".").canonicalize().unwrap();
+        if folder.to_str().unwrap().find("dove").is_none() {
+            folder.push("dove");
+        }
+        folder
+    };
+    // Project name and path
+    let project_name = "demoproject_8";
+    let project_folder = {
+        let mut folder = dove_folder.clone();
+        folder.push(project_name);
+        if folder.exists() {
+            remove_dir_all(&folder).expect(&format!(
+                "[ERROR] Couldn't delete project directory: {}",
+                folder.to_str().unwrap()
+            ));
+        }
+        folder
+    };
+
+    // $ cargo run -- new demoproject_6 -d pont
+    // $ dove new demoproject_6 -d pont
+    {
+        let mut dove_new = Command::new("cargo");
+        dove_new
+            .args(&["run", "--", "new", project_name])
+            .args(&["-d", "pont"])
+            .current_dir(&dove_folder);
+        let command_string = format!("{:?} ", dove_new).replace("\"", "");
+        let result = dove_new
+            .output()
+            .expect(&format!("[RUN]: {}", command_string));
+        let code = result.status.code().unwrap();
+        assert_eq!(
+            0,
+            code,
+            "[ERROR] Command: {}\r\nCode: {}\r\nMessage: {}\r\n",
+            command_string,
+            code,
+            String::from_utf8(result.stderr).unwrap()
+        );
+        set_dependencies_local_move_stdlib(&project_folder);
+    }
+
+    // $ cargo run -- build
+    // $ dove build
+    {
+        let mut dove_build = Command::new("cargo");
+        dove_build
+            .args(&["run", "--", "build"])
+            .current_dir(&project_folder);
+        let command_string = format!("{:?} ", dove_build).replace("\"", "");
+        let result = dove_build
+            .output()
+            .expect(&format!("[RUN]: {}", command_string));
+        let code = result.status.code().unwrap();
+        assert_eq!(
+            0,
+            code,
+            "[ERROR] Command: {}\r\nCode: {}\r\nMessage: {}\r\n",
+            command_string,
+            code,
+            String::from_utf8(result.stderr).unwrap()
+        );
+    }
+
+    // project_folder/scripts/demo.move
+    {
+        let script_path = {
+            let mut path = project_folder.clone();
+            path.push("scripts");
+            path.push("rdemo.move");
+            path
+        };
+        write_all(
+            &script_path,
+            "script {
+                    fun main(x:u64,y:u64) {
+                        let _result = x + y;
+                    }
+                }",
+        )
+        .unwrap();
+    }
+    // $ cargo run -- run rdemo.move -a 3 5
+    // $ dove run rdemo.move
+    {
+        let mut dove_run = Command::new("cargo");
+        dove_run
+            .args(&["run", "--"])
+            .args(&["run", "rdemo.move"])
+            .args(&["-a", "3", "5"])
+            .current_dir(&project_folder);
+        let command_string = format!("{:?} ", dove_run).replace("\"", "");
+        let result = dove_run
+            .output()
+            .expect(&format!("[RUN]: {}", command_string));
+        let code = result.status.code().unwrap();
+        assert_eq!(
+            0,
+            code,
+            "[ERROR] Command: {}\r\nCode: {}\r\nMessage: {}\r\n",
+            command_string,
+            code,
+            String::from_utf8(result.stderr.clone()).unwrap()
+        );
+        let stdout = String::from_utf8(result.stdout).unwrap();
+        assert!(
+            stdout.contains("main ...... ok"),
+            "[ERROR] Command: {}. \r\nMessage: {}",
+            command_string,
+            stdout
+        );
+    }
+
+    remove_dir_all(&project_folder).expect(&format!(
+        "[ERROR] Couldn't delete project directory: {}",
+        project_folder.to_str().unwrap()
+    ));
+}
+
+fn set_dependencies_local_move_stdlib(project_path: &PathBuf) {
+    use toml::Value;
+
+    let mut dove_toml_path = project_path.clone();
+    dove_toml_path.push("Dove.toml");
+    let mut toml_value = read_to_string(&dove_toml_path)
+        .unwrap()
+        .parse::<Value>()
+        .unwrap();
+    {
+        let v = toml_value
+            .get_mut("package")
+            .unwrap()
+            .get_mut("dependencies")
+            .unwrap()
+            .as_array_mut()
+            .unwrap();
+        v.clear();
+        let mut dd = toml::map::Map::new();
+        dd.insert(
+            "path".to_string(),
+            Value::String("../tests/move-stdlib".to_string()),
+        );
+        v.push(Value::Table(dd));
+    }
+    write_all(
+        &dove_toml_path,
+        toml::to_string(&toml_value).unwrap().as_str(),
+    )
+    .unwrap();
+}
